+++ src/heatmap-component/class-names.ts
@@ -1,0 +1,17 @@
+/** Class names of DOM elements */
+export const Class = {
+    MainDiv: 'heatmap-main-div',
+    CanvasDiv: 'heatmap-canvas-div',
+    Marker: 'heatmap-marker',
+    MarkerX: 'heatmap-marker-x',
+    MarkerY: 'heatmap-marker-y',
+    TooltipBox: 'heatmap-tooltip-box',
+    TooltipContent: 'heatmap-tooltip-content',
+    PinnedTooltipBox: 'heatmap-pinned-tooltip-box',
+    PinnedTooltipContent: 'heatmap-pinned-tooltip-content',
+    PinnedTooltipPin: 'heatmap-pinned-tooltip-pin',
+    PinnedTooltipClose: 'heatmap-pinned-tooltip-close',
+    Overlay: 'heatmap-overlay',
+    OverlayShade: 'heatmap-overlay-shade',
+    OverlayMessage: 'heatmap-overlay-message',
+} as const;