+++ src/heatmap-component/heatmap-core.ts
@@ -1,7 +1,7 @@
 import { Class } from './class-names';
 import * as d3 from './d3-modules';
 import { DataDescription } from './data/data-description';
-import { Behavior, HotmapExtension } from './extension';
+import { Behavior, HeatmapExtension } from './extension';
 import { Scales } from './scales';
 import { State } from './state';
 import { attrd } from './utils';
@@ -10,7 +10,7 @@ import { attrd } from './utils';
 export class HeatmapCore<TX, TY, TItem> {
     protected readonly state: State<TX, TY, TItem>;
 
-    registerExtension<TParams extends {}, TDefaults extends TParams>(extension: HotmapExtension<TParams, TDefaults>, params?: Partial<TParams>): Behavior<TParams> {
+    registerExtension<TParams extends {}, TDefaults extends TParams>(extension: HeatmapExtension<TParams, TDefaults>, params?: Partial<TParams>): Behavior<TParams> {
         const behavior = extension.create(this.state, params);
         behavior.register();
         return behavior;
@@ -32,7 +32,7 @@ export class HeatmapCore<TX, TY, TItem> {
             console.error(`This ${this.constructor.name} has already been rendered in element`, this.state.dom.rootDiv.node());
             throw new Error(`This ${this.constructor.name} has already been rendered. Cannot render again.`);
         }
-        console.time('Hotmap render');
+        console.time('Heatmap render');
 
         const rootDiv: d3.Selection<HTMLDivElement, any, any, any> = (typeof divElementOrId === 'string') ? d3.select(`#${divElementOrId}`) : d3.select(divElementOrId);
         if (rootDiv.empty()) throw new Error('Failed to initialize, wrong div ID?');
@@ -59,15 +59,15 @@ export class HeatmapCore<TX, TY, TItem> {
         });
         this.state.dom = { rootDiv, mainDiv, canvasDiv, canvas, svg };
 
-        svg.on('mousemove.hotmapcore', (e: MouseEvent) => this.state.events.hover.next(this.state.getPointedItem(e)));
-        svg.on('mouseleave.hotmapcore', (e: MouseEvent) => this.state.events.hover.next(undefined));
-        svg.on('click.hotmapcore', (e: MouseEvent) => this.state.events.select.next(this.state.getPointedItem(e)));
+        svg.on('mousemove.heatmapcore', (e: MouseEvent) => this.state.events.hover.next(this.state.getPointedItem(e)));
+        svg.on('mouseleave.heatmapcore', (e: MouseEvent) => this.state.events.hover.next(undefined));
+        svg.on('click.heatmapcore', (e: MouseEvent) => this.state.events.select.next(this.state.getPointedItem(e)));
 
         this.state.events.render.next(undefined);
         this.state.emitResize();
-        d3.select(window).on('resize.resizehotmapcanvas', () => this.state.emitResize());
+        d3.select(window).on('resize.resizeheatmapcanvas', () => this.state.emitResize());
 
-        console.timeEnd('Hotmap render');
+        console.timeEnd('Heatmap render');
         return this;
     }