+++ src/main.ts
@@ -1,4 +1,4 @@
-/** Main file for importing `Hotmap` as a dependency */
+/** Main file for importing `HeatmapComponent` as a dependency */
 
-export { demo, demo2, demo3 } from './hotmap/demo';
-export { Heatmap } from './hotmap/heatmap';
+export { demo, demo2, demo3 } from './heatmap-component/demo';
+export { Heatmap } from './heatmap-component/heatmap';