+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+-   Renamed the package from `hotmap` to `heatmap-component`
 -   Zooming is off by default (call `hm.setZooming({ axis: 'x' })` to turn it on)
 
 ## [0.8.0] - 2024-04-26